services/ner-service/pyproject.toml
[project]
name = "ner-service"
version = "0.1.0"
description = "Medical NER service for the Clinical AI Platform (stubbed for Milestone 0)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn[standard]>=0.30.0,<0.31.0",
    "pydantic>=2.8.0,<3.0.0",
    "scispacy==0.5.4",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*", "services*"]

